requirements.txt
Flask==3.1.2
PyMySQL==1.1.2
python-dotenv==1.1.1
argon2-cffi==25.1.0
bcrypt==5.0.0